requirements.txt
opencv-python
numpy
torch
torchvision
waymo-open-dataset-tf-2-11-0
matplotlib
Pillow
easydict
tqdm